Moved metrics to a separate file
